requirements.txt
azure-core==1.32.0
azure-storage-blob==12.24.0
bleach==6.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
greenlet==3.1.1
idna==3.10
isodate==0.7.2
kaggle==1.6.17
numpy==2.2.1
pandas==2.2.3
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
